+++ config.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import psutil
 import uuid
@@ -22,7 +23,10 @@ class CreateConfig():
             os.environ[env] = "False"
             return False
 
-    # not error so far...
+    # Catch Flask logger.
+    logger = logging.getLogger("flaskapp")
+
+    # No error found so far...
     FLASK_INIT_ERROR = ""
 
     # FLASK_ENV
@@ -31,13 +35,13 @@ class CreateConfig():
 
     # FLASK_RESULTS_ARE_PUBLIC
     FLASK_RESULTS_ARE_PUBLIC = format_true_false("FLASK_RESULTS_ARE_PUBLIC",
-                                                 "False")
-    print("FLASK_RESULTS_ARE_PUBLIC is", FLASK_RESULTS_ARE_PUBLIC)
+                                                 "True")
+    logger.info(f"FLASK_RESULTS_ARE_PUBLIC: {FLASK_RESULTS_ARE_PUBLIC}")
 
     # FLASK_RESULTS_BY_EMAIL
     FLASK_RESULTS_BY_EMAIL = format_true_false("FLASK_RESULTS_BY_EMAIL",
                                                "False")
-    print("FLASK_RESULTS_BY_EMAIL is", FLASK_RESULTS_BY_EMAIL)
+    logger.info(f"FLASK_RESULTS_BY_EMAIL: {FLASK_RESULTS_BY_EMAIL}")
 
     if FLASK_RESULTS_ARE_PUBLIC == False \
       and FLASK_RESULTS_BY_EMAIL == False:
@@ -63,7 +67,7 @@ class CreateConfig():
     if FLASK_RESULTS_BY_EMAIL:
         MAIL_USE_TLS = format_true_false("MAIL_USE_TLS",
                                          "False")
-        print("MAIL_USE_TLS:", MAIL_USE_TLS)
+        logger.info(f"MAIL_USE_TLS: {MAIL_USE_TLS}")
 
     # FLASK_MAX_JOBS
     if ("FLASK_MAX_JOBS" not in os.environ) \
@@ -71,7 +75,7 @@ class CreateConfig():
         FLASK_MAX_JOBS = psutil.cpu_count() - 1
     else:
         FLASK_MAX_JOBS = int(os.environ["FLASK_MAX_JOBS"])
-    print("FLASK_MAX_JOBS:", FLASK_MAX_JOBS)
+    logger.info(f"FLASK_MAX_JOBS: {FLASK_MAX_JOBS}")
 
     # FLASK_JOB_TIMEOUT
     if ("FLASK_JOB_TIMEOUT" not in os.environ) \
@@ -79,21 +83,23 @@ class CreateConfig():
         FLASK_JOB_TIMEOUT = 24
     else:
         FLASK_JOB_TIMEOUT = int(os.environ["FLASK_JOB_TIMEOUT"])
-    print("FLASK_JOB_TIMEOUT:", FLASK_JOB_TIMEOUT)
+    logger.info(f"FLASK_JOB_TIMEOUT: {FLASK_JOB_TIMEOUT}")
 
 
     # Search AutoClass C executable
-    autoclass_version = wrapper.get_autoclass_version()
-    if not autoclass_version:
+    AUTOCLASSC_VERSION = wrapper.get_autoclass_version()
+    AUTOCLASSC_PATH = wrapper.search_autoclass_in_path()
+    if not AUTOCLASSC_VERSION:
         FLASK_INIT_ERROR = "Cannot find/run AutoClass C executable."
 
     # Print and store autoclasswrapper version
-    print("autoclasswrapper version: ", wrapper.__version__)
-    WRAPPER_VERSION = wrapper.__version__
+    AUTOCLASSWRAPPER_VERSION = wrapper.__version__
+    logger.info(f"AutoClassWrapper version: {AUTOCLASSWRAPPER_VERSION}")
+    
 
     # internal parameters
     if "SECRET_KEY" not in os.environ:
-        print("No SECRET_KEY found in env. Generating.")
+        logger.info("No SECRET_KEY found in env. Generating.")
     SECRET_KEY = os.environ.get("SECRET_KEY", str(uuid.uuid4()))
     RESULTS_FOLDER = os.path.join(os.environ["FLASK_HOME"], "results")
     JOB_NAME_LENGTH = 6