+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Use pathlib in `export_results.py`
 - Use conda env instead of pipenv
 - Move config file to the config/ directory
+- Use logging for messages
 
 **0.1.3**
 - Fix bug for discrete date type